requirements.txt
pygame-ce
numpy
scipy
fastdtw